requirements.txt
langchain
langchain-community
chromadb
openai
tiktoken
streamlit
pandas
python-dotenv 